refactor: extract url helper functions into own module
