+++ sso-url-helper/build.gradle
@@ -1,6 +1,5 @@
 apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
 apply plugin: 'maven-publish' // needed for JitPack.io
 
 android {
@@ -44,7 +43,7 @@ afterEvaluate {
                 from components.release
 
                 groupId = 'com.github.stefan-niedermann.nextcloud-commons'
-                artifactId = 'sso-glide'
+                artifactId = 'sso-url-helper'
             }
         }
     }
@@ -57,13 +56,7 @@ dependencies {
 
     api "com.github.nextcloud:Android-SingleSignOn:$sso_version"
 
-    implementation project(':sso-url-helper')
-
-    implementation 'com.github.bumptech.glide:glide:4.16.0'
-    kapt 'com.github.bumptech.glide:compiler:4.16.0'
-
     testImplementation 'androidx.test:core:1.5.0'
     testImplementation "io.mockk:mockk:1.13.7"
     testImplementation 'junit:junit:4.13.2'
-    testImplementation 'org.robolectric:robolectric:4.11.1'
 }
