+++ sso-url-helper/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="it.niedermann.nextcloud.sso.url.helper" />
